+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 4.15.0 (2025-04-20)
+
+#### Improvements
+
+  - Added the `pgtrigger.Composer` trigger utility for writing conditional statement-level triggers by [@wesleykendall](https://github.com/wesleykendall) in [#204](https://github.com/AmbitionEng/django-pgtrigger/pull/204).
+  - Provide statement-level versions of `pgtrigger.Protect` and `pgtrigger.ReadOnly` by [@wesleykendall](https://github.com/wesleykendall) in [#204](https://github.com/AmbitionEng/django-pgtrigger/pull/204).
+
+#### Fixes
+
+  - Fixes an assertion error that can happen during migrations by [@wesleykendall](https://github.com/wesleykendall) in [#204](https://github.com/AmbitionEng/django-pgtrigger/pull/204).
+
 ## 4.14.0 (2025-04-18)
 
 #### Changes